+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Method to check if a string is a valid IP V4 ([#2](https://github.com/khalyomede/ip/issues/2)).
+- Method to check if a string is a valid IP V6 ([#2](https://github.com/khalyomede/ip/issues/2)).
+
 ## [0.2.0] - 2025-10-21
 
 ### Added